Initial source release. This is work in progress and not generally useful, yet. The most notable short comings are lack of support for multiple sandboxed threads and lack of support of signals.
Also, for debugging purposes, the sandbox current defaults to allowing all
system calls. And it does not yet implement all security critical code in
assembly. As such, it is vulnerable to attacks from malicious threads that
modify shared memory.

Finally, we are missing any documentation on the mode of operation. These
parts will be added as the project progresses.


git-svn-id: http://seccompsandbox.googlecode.com/svn/trunk@2 55e79e8e-603c-11de-8c10-5fe6993ea61f