Added support for signal handling in the sandbox. In the first version, this now gives us access to sigprocmask() and rt_sigprocmask().
But it also means, we had to tighten the code that prevented an attack on the
trusted thread by means of making it execute a signal handler. We now ensure
that our stack pointer is always invalid, while running outside of Seccomp mode.

For now, we don't allow sigaltstack(), as that could break this security
mechanism. Ultimately, I believe we can add support for sigaltstack(), but have
to be very careful how we do this.


git-svn-id: http://seccompsandbox.googlecode.com/svn/trunk@70 55e79e8e-603c-11de-8c10-5fe6993ea61f