Add test to check that clone() preserves registers
This is a bit of a hassle, because it requires an assembly code
helper, but it is an interesting exercise.  In order to test this, we
dump registers into global variables.  I have split out some of
test_clone into separate functions.

It is useful because current glibc versions do not rely on most of the
registers being preserved (so this is otherwise untested), but glibc
could rely on this in future.

Also I would like to simplify the register preservation code later,
and I want a test that ensures I don't break it.

Review URL: http://codereview.chromium.org/3443010

git-svn-id: http://seccompsandbox.googlecode.com/svn/trunk@121 55e79e8e-603c-11de-8c10-5fe6993ea61f