clone(): Copy registers across in untrusted code instead of trusted code
Copy register values to restore into the signal stack frame early on,
in sandbox_clone() (untrusted code).  This means we don't need to
include the register values in the message sent to the trusted
process, and we don't need to copy them to and from the secure memory
area.

In the trusted thread code, we no longer need to store the
return-address-or-zero in %mm3 (i386) or %r15 (x86-64).  The return
address is always in the signal stack frame, so we can remove two
conditionals for the special case of the first thread.

Review URL: http://codereview.chromium.org/3394009

git-svn-id: http://seccompsandbox.googlecode.com/svn/trunk@122 55e79e8e-603c-11de-8c10-5fe6993ea61f