Fix hole that allows tampering with sendmsg() arguments
Ensure that "struct msghdr" is stored in a page that is mapped as
MAP_PRIVATE by reserving space in the executable/library's data
segment, which will be a protected mapping.

Before, the trusted thread used the stack provided by untrusted code,
which could have been mapped with MAP_SHARED and so could be shared
between the trusted fork()'d helper process and the untrusted threads.

BUG=http://code.google.com/p/seccompsandbox/issues/detail?id=3
TEST=None: The bug involves a race condition which I haven't tried
  to exploit.

Review URL: http://codereview.chromium.org/3532002

git-svn-id: http://seccompsandbox.googlecode.com/svn/trunk@124 55e79e8e-603c-11de-8c10-5fe6993ea61f