Finally eliminated the dirty hack that we used for populating the system call table. This table needs to be in read-only memory, and it needs to have all of the unused entries zero'd out. We used to rely a C having designated initializers. And we also abused a gcc extension to move the data structure into read-only memory.
Instead, we are now explicitly managing this data structure, which allows us
to remove our dependency on C. All code is now either C++ or assembly. This
had a trickle down effect of cleaning up a lot of other code and getting rid
of a lot of "asm" labels for identifiers that are no longer referenced by
anything but C++ code.

BUG=http://code.google.com/p/chromium/issues/detail?id=88578
TEST=make test
Review URL: http://codereview.chromium.org/7397023

git-svn-id: http://seccompsandbox.googlecode.com/svn/trunk@160 55e79e8e-603c-11de-8c10-5fe6993ea61f