Implement TLS initialisation system calls for i386 and x86-64
This will allow ld.so (the dynamic linker) to start up under the
sandbox.

This is not straightforward because the set_thread_area()/arch_prctl()
syscalls modify a thread's state, so we cannot simply forward the
syscall to the trusted thread, because that would modify the wrong
thread's state.

Instead, we use the trick of using clone() to create a new thread that
picks up where the original thread left off, but with modified thread
state.  Furthermore, clone() provides a built-in way of setting up
TLS.

This requires two further tricks:
 * Capture a signal frame that we can use to return into the new thread.
 * Do write() and then _exit() without using the stack inbetween,
   because the stack is used by the new thread that write() spawns.

Pleasantly, this can all be done without changing the trusted code.

BUG=http://code.google.com/p/seccompsandbox/issues/detail?id=13
TEST=test_tls_setup_via_arch_prctl, test_tls_setup_via_set_thread_area

Review URL: http://codereview.chromium.org/4848001

git-svn-id: http://seccompsandbox.googlecode.com/svn/trunk@162 55e79e8e-603c-11de-8c10-5fe6993ea61f